rand::distributions -> distr; split uniform module (#1470)
