+++ benches/Cargo.toml
@@ -15,8 +15,8 @@ criterion = "0.5"
 criterion-cycles-per-byte = "0.6"
 
 [[bench]]
-name = "distributions"
-path = "src/distributions.rs"
+name = "distr"
+path = "src/distr.rs"
 harness = false
 
 [[bench]]
