+++ benches/benches/misc.rs
@@ -14,7 +14,7 @@ const RAND_BENCH_N: u64 = 1000;
 
 use test::Bencher;
 
-use rand::distributions::{Bernoulli, Distribution, Standard};
+use rand::distr::{Bernoulli, Distribution, Standard};
 use rand::prelude::*;
 use rand_pcg::{Pcg32, Pcg64Mcg};
